+++ AlmacenApi.Application/Features/Combo/Dto/ProductComboResultDto.cs
@@ -3,6 +3,6 @@ public class ProductcomboResultDto
 {
     public required string? id {get ; set ;}
     public required string? NameProduct {get ; set ;}
-    public required int Quantity {get ; set ;}
+    public required double Quantity {get ; set ;}
     public required string? Unity {get; set ;} = string.Empty;
 }
