+++ AlmacenApi.Application/Features/ComboOut/ComboOutDto.cs
@@ -2,5 +2,5 @@ namespace AlmacenApi.Aplication.Features.ComboOut;
 public class ComboOutDto
 {
     public required string ProductDto {get ; set ;}
-    public required int Quantity{get ; set ;}
+    public required double Quantity{get ; set ;}
 }
